pyproject.toml
[project]
name = "longzhongxian-recruit"
version = "0.1.0"
description = "笼中仙 AI 招聘助手 - Boss直聘线上初筛"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "structlog>=24.0",
    "alembic>=1.13",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio>=0.23",
    "aiosqlite>=0.20",
    "types-PyYAML>=6.0",
]

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]

[tool.mypy]
python_version = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
